requirements.txt
Flask==0.10.1
Flask-Bcrypt==0.6.2
Flask-Login==0.2.11
Flask-RESTful==0.3.2
Flask-Script==2.0.5
Flask-WTF==0.11
Jinja2==2.7.3
MarkupSafe==0.23
Unidecode==0.04.17
WTForms==2.0.2
Werkzeug==0.10.4
aniso8601==0.92
argparse==1.2.1
awesome-slugify==1.6.4
flask-mongoengine==0.7.1
gunicorn==19.3.0
itsdangerous==0.24
mongoengine==0.9.0
pymongo==2.8
python-bcrypt==0.3.1
pytz==2015.2
regex==2015.03.18
six==1.9.0
wsgiref==0.1.2
